Chore:: refactor web3 donation panel, getdonation Adress
